+++ masterwork/campaign.py
@@ -25,7 +25,7 @@ import os
 import statistics
 import sys
 
-from pipeline import line as line_mod
+from masterwork import line as line_mod
 
 PUAN_KEYS = ("axes",)